Do not overwrite tags specified in the config file
The issue was that tags specified in config were overriden in the
`Spinach::Cli` when no tag were set using `-t` or `--tags`. I kept the old
behaviour of adding the `~wip` tag if no tag was defined or it is not specified.
But probably it should only be done when no tag is set (either in config or
command line), because if you want to run only scenarios with `@wip` tag then
you're relying on the fact that a match tag have precedence over a negated tag.