fix test failure only present with some library versions
I could narrow it down to the exact versions of which librariess and
figure out exactly why that combination made the failure
happen. However, I can't find a compelling reason why this test
shouldn't pass in step_definitions like the others in this file, and
that change makes the test pass with the library version combos where
it previously failed, and doesn't affect it elsewhere. So, we're going
with that.