Finally randomize feature & scenario ordering (#204)
* fix 2 documentation mistakes

* removing extraneous whitespace

* Stable identifiers for ordering features and scenarios.

We need the relative filename involved in both so that a seed can
generate stable randomizations across runs and machines. Similarly,
that's why we need the line number for Scenario#ordering_id.

* introduce Orderers without hooking them up to anything

* hook up Orderers to Reporters. still not reordering tests at all

* Add features for testing randomization and random seeding

* Add cli tests for --seed and --rand

* Finally (re)order features/scenarios according to orderer instances

* Randomize our own test runs.

* fix test failure only present with some library versions

I could narrow it down to the exact versions of which librariess and
figure out exactly why that combination made the failure
happen. However, I can't find a compelling reason why this test
shouldn't pass in step_definitions like the others in this file, and
that change makes the test pass with the library version combos where
it previously failed, and doesn't affect it elsewhere. So, we're going
with that.